+++ lib/data/repositories/todos_repository/models/todo.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart' as m;
 import 'package:meta/meta.dart';
-import 'package:todos/repositories/todos_repository/models/todo_priority.dart';
-import 'package:todos/repositories/todos_repository/models/theme.dart';
+import 'package:todos/data/repositories/todos_repository/models/theme.dart';
+import 'package:todos/data/repositories/todos_repository/models/todo_priority.dart';
 import 'package:uuid/uuid.dart';
 
 class Todo {
