+++ lib/data/repositories/todos_repository/repositories/mock_todos_repository.dart
@@ -1,8 +1,8 @@
-import 'package:todos/repositories/todos_repository/models/branch.dart';
-import 'package:todos/repositories/todos_repository/models/image.dart';
-import 'package:todos/repositories/todos_repository/models/step.dart';
-import 'package:todos/repositories/todos_repository/models/todo.dart';
-import 'package:todos/repositories/todos_repository/repositories/i_todos_repository.dart';
+import 'package:todos/data/repositories/todos_repository/models/branch.dart';
+import 'package:todos/data/repositories/todos_repository/models/image.dart';
+import 'package:todos/data/repositories/todos_repository/models/step.dart';
+import 'package:todos/data/repositories/todos_repository/models/todo.dart';
+import 'package:todos/data/repositories/todos_repository/repositories/i_todos_repository.dart';
 import 'package:uuid/uuid.dart';
 
 class MockTodosRepository implements ITodosRepository {
